add: index tests
