+++ src/frontend/src/store/modules/index.js
@@ -2,10 +2,12 @@ import Auth from "@/store/modules/auth.store";
 import Builder from "@/store/modules/builder.store";
 import Cart from "@/store/modules/cart.store";
 import Orders from "@/store/modules/orders.store";
+import Common from "@/store/modules/common.store";
 
 export default {
   Auth,
   Builder,
   Cart,
   Orders,
+  Common,
 };
